pyproject.toml
[project]
name = "introspect"
version = "0.1.0"
description = "Introspection experiments with steering vectors in language models"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.5.0",
    "transformers>=4.48.0",
    "accelerate>=1.3.0",
    "tqdm>=4.67.0",
    "requests>=2.32.0",
    "pyyaml>=6.0",
]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu130" }

[project.optional-dependencies]
dev = [
    "ipython>=8.30.0",
    "jupyter>=1.1.0",
]

[dependency-groups]
dev = [
    "ipython>=8.30.0",
]
